first beta release

+++ tools/game_macros_hid_map.h
@@ -9,6 +9,6 @@ typedef struct {
 } GameMacrosMap;
 
 uint16_t game_macros_get_hid_pair(const char* key);
-uint32_t game_macros_parse_number(const char* key);
+size_t game_macros_parse_number(const char* key);
 size_t game_macros_string_to_int(const char* key);
 extern GameMacrosMap game_macros_hid_pairs[];
